+++ .gitignore
@@ -143,6 +143,7 @@ venv/
 ENV/
 env.bak/
 venv.bak/
+.zshrc
 
 # Spyder project settings
 .spyderproject
