+++ app/account/tests/test_user_api.py
@@ -9,9 +9,9 @@ from rest_framework.test import APIClient
 from rest_framework import status
 
 
-CREATE_USER_URL = reverse('user:create')
-TOKEN_URL = reverse('user:token')
-ME_URL = reverse('user:me')
+CREATE_USER_URL = reverse('account:create')
+TOKEN_URL = reverse('account:token')
+ME_URL = reverse('account:me')
 
 
 def create_user(**params):
