+++ app/account/views.py
@@ -5,7 +5,7 @@ from rest_framework import generics, authentication, permissions
 from rest_framework.authtoken.views import ObtainAuthToken
 from rest_framework.settings import api_settings
 
-from user.serializers import (
+from account.serializers import (
     UserSerializer,
     AuthTokenSerializer,
 )
