+++ app/app/urls.py
@@ -29,5 +29,5 @@ urlpatterns = [
         name="api-docs",
     ),
 
-    path("api/user/", include("user.urls")),
+    path("api/account/", include("account.urls")),
 ]
