+++ app/core/admin.py
@@ -5,7 +5,7 @@ from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin as BaseUserAdmin
 from django.utils.translation import gettext_lazy as _
 
-from core import models
+from account import models
 
 
 class UserAdmin(BaseUserAdmin):
